Fix enum deletion after object deletion in migration v2
