+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-migration-runner-v2.module.ts
@@ -12,10 +12,10 @@ import { WorkspaceMetadataIndexActionRunnerService } from 'src/engine/workspace-
 import { WorkspaceMetadataMigrationRunnerService } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-metadata-migration-runner/workspace-metadata-migration-runner-service';
 import { WorkspaceMetadataObjectActionRunnerService } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-metadata-migration-runner/workspace-metadata-object-action-runner.service';
 import { WorkspaceMigrationRunnerV2Service } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-migration-runner-v2.service';
-import { WorkspaceSchemaFieldActionRunnerService } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/workspace-schema-field-action-runner.service';
-import { WorkspaceSchemaIndexActionRunnerService } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/workspace-schema-index-action-runner.service';
+import { WorkspaceSchemaFieldActionRunnerModule } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/workspace-schema-field-action-runner/workspace-schema-field-action-runner.module';
+import { WorkspaceSchemaIndexActionRunnerService } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/workspace-schema-index-action-runner/workspace-schema-index-action-runner.service';
 import { WorkspaceSchemaMigrationRunnerService } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/workspace-schema-migration-runner.service';
-import { WorkspaceSchemaObjectActionRunnerService } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/workspace-schema-object-action-runner.service';
+import { WorkspaceSchemaObjectActionRunnerService } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/workspace-schema-object-action-runner/workspace-schema-object-action-runner.service';
 
 @Module({
   imports: [
@@ -26,6 +26,7 @@ import { WorkspaceSchemaObjectActionRunnerService } from 'src/engine/workspace-m
     WorkspacePermissionsCacheModule,
     WorkspaceMetadataCacheModule,
     WorkspaceSchemaManagerModule,
+    WorkspaceSchemaFieldActionRunnerModule,
   ],
   providers: [
     WorkspaceMetadataObjectActionRunnerService,
@@ -33,21 +34,10 @@ import { WorkspaceSchemaObjectActionRunnerService } from 'src/engine/workspace-m
     WorkspaceMetadataFieldActionRunnerService,
     WorkspaceSchemaObjectActionRunnerService,
     WorkspaceSchemaIndexActionRunnerService,
-    WorkspaceSchemaFieldActionRunnerService,
     WorkspaceMetadataMigrationRunnerService,
     WorkspaceSchemaMigrationRunnerService,
     WorkspaceMigrationRunnerV2Service,
   ],
-  exports: [
-    WorkspaceMigrationRunnerV2Service,
-    WorkspaceMetadataObjectActionRunnerService,
-    WorkspaceMetadataIndexActionRunnerService,
-    WorkspaceMetadataFieldActionRunnerService,
-    WorkspaceSchemaObjectActionRunnerService,
-    WorkspaceSchemaIndexActionRunnerService,
-    WorkspaceSchemaFieldActionRunnerService,
-    WorkspaceMetadataMigrationRunnerService,
-    WorkspaceSchemaMigrationRunnerService,
-  ],
+  exports: [WorkspaceMigrationRunnerV2Service],
 })
 export class WorkspaceMigrationRunnerV2Module {}
