+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/workspace-schema-field-action-runner/actions/workspace-schema-field-create-action-runner.service.ts
@@ -1,0 +1,60 @@
+import { Injectable } from '@nestjs/common';
+
+import { WorkspaceSchemaManagerService } from 'src/engine/twenty-orm/workspace-schema-manager/workspace-schema-manager.service';
+import { CreateFieldAction } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/workspace-migration-field-action-v2';
+import { WorkspaceMigrationActionRunnerArgs } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/types/workspace-migration-action-runner-args.type';
+import { generateColumnDefinitions } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/utils/generate-column-definitions.util';
+import { prepareWorkspaceSchemaContext } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/utils/workspace-schema-context.util';
+import {
+  collectEnumOperationsForField,
+  EnumOperation,
+  executeBatchEnumOperations,
+} from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/utils/workspace-schema-enum-operations.util';
+
+@Injectable()
+export class WorkspaceSchemaFieldCreateActionRunnerService {
+  constructor(
+    private readonly workspaceSchemaManagerService: WorkspaceSchemaManagerService,
+  ) {}
+
+  public async run({
+    action,
+    queryRunner,
+    flatObjectMetadataMaps,
+  }: WorkspaceMigrationActionRunnerArgs<CreateFieldAction>): Promise<void> {
+    const { flatFieldMetadata } = action;
+    const {
+      schemaName,
+      tableName,
+      flatObjectMetadataWithFlatFieldMaps: flatObjectMetadata,
+    } = prepareWorkspaceSchemaContext({
+      flatObjectMetadataMaps,
+      objectMetadataId: flatFieldMetadata.objectMetadataId,
+    });
+
+    const enumOperations = collectEnumOperationsForField({
+      flatFieldMetadata: flatFieldMetadata,
+      tableName,
+      operation: EnumOperation.CREATE,
+    });
+
+    await executeBatchEnumOperations({
+      enumOperations,
+      queryRunner,
+      schemaName,
+      workspaceSchemaManagerService: this.workspaceSchemaManagerService,
+    });
+
+    const columnDefinitions = generateColumnDefinitions({
+      flatFieldMetadata: flatFieldMetadata,
+      flatObjectMetadataWithoutFields: flatObjectMetadata,
+    });
+
+    await this.workspaceSchemaManagerService.columnManager.addColumns({
+      queryRunner,
+      schemaName,
+      tableName,
+      columnDefinitions,
+    });
+  }
+}
