+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/workspace-schema-object-action-runner/workspace-schema-object-action-runner.service.ts
@@ -15,13 +15,12 @@ import {
 import { type RunnerMethodForActionType } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/types/runner-method-for-action-type';
 import { type WorkspaceMigrationActionRunnerArgs } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/types/workspace-migration-action-runner-args.type';
 import { generateColumnDefinitions } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/utils/generate-column-definitions.util';
-
-import { prepareWorkspaceSchemaContext } from './utils/workspace-schema-context.util';
+import { prepareWorkspaceSchemaContext } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/utils/workspace-schema-context.util';
 import {
   collectEnumOperationsForObject,
   EnumOperation,
   executeBatchEnumOperations,
-} from './utils/workspace-schema-enum-operations.util';
+} from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/utils/workspace-schema-enum-operations.util';
 
 @Injectable()
 export class WorkspaceSchemaObjectActionRunnerService
@@ -49,20 +48,23 @@ export class WorkspaceSchemaObjectActionRunnerService
       tableName,
     });
 
-    const enumFlatFieldMetadatas = Object.values(
+    const enumOrCompositeFlatFieldMetadatas = Object.values(
       flatObjectMetadataWithFlatFieldMaps.fieldsById,
     )
       .filter((field): field is FlatFieldMetadata => field != null)
-      .filter((field) => isEnumFlatFieldMetadata(field));
+      .filter(
+        (field) =>
+          isEnumFlatFieldMetadata(field) || isCompositeFlatFieldMetadata(field),
+      );
 
     const enumOperations = collectEnumOperationsForObject({
-      flatFieldMetadatas: enumFlatFieldMetadatas,
+      flatFieldMetadatas: enumOrCompositeFlatFieldMetadatas,
       tableName,
       operation: EnumOperation.DROP,
     });
 
     await executeBatchEnumOperations({
-      enumOperations,
+      enumOperations: enumOperations,
       queryRunner,
       schemaName,
       workspaceSchemaManagerService: this.workspaceSchemaManagerService,
